# api.yaml
swagger: "2.0"
info:
  title: Learn Chinese Api
  description: APIs for learning chinese
  version: 1
consumes:
  - application/json
produces:
  - application/json
paths:
  /addWord/detailed:
    post:
      summary: Path for adding a new word and all its details
      tags:
      - addWord
      parameters:
      - name: body
        in: body
        description: the data sent from front end
        required: true
        schema:
          type: array
          example: [{
            "id":"3",
            "pingYing":"sān",
            "meaning":"three",
            "character":"三",
            "examples":"三个",
            "notes":""}]
      responses:
        "200":
          description: Success
  /addWord/simple:
    post:
      summary: Path for inputting a word and then fetch meta data via the Line website
      tags:
      - search
      parameters:
      - name: body
        in: body
        description: the data sent from front end
        required: true
        schema:
          type: object
          properties:
            id:
              type: string
            character:
              type: string
            notes:
              type: string
          example: {
            "id":"3",
            "character":"三",
            "notes": ""}
      responses:
        "200":
          description: Success
...